+++ app/src/main/java/com/example/yomikaze_app_kotlin/Presentation/Components/ComicCard/RankingComicCard/NormalComicCard.kt
@@ -31,13 +31,13 @@ import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import coil.request.CachePolicy
 import coil.request.ImageRequest
-import com.example.yomikaze_app_kotlin.Presentation.Components.ComicCard.ShareComponents.TagComponent
 import com.example.yomikaze_app_kotlin.Presentation.Components.ComicCard.ShareComponents.IconAndNumbers
+import com.example.yomikaze_app_kotlin.Presentation.Components.ComicCard.ShareComponents.TagComponent
 import com.example.yomikaze_app_kotlin.R
 
 @Composable
 fun NormalComicCard(
-    comicId: Int,
+    comicId: Long,
     image: String,
     comicName: String,
     status: String,
